feat: Add likesReducer
