+++ src/contexts/index.js
@@ -1,2 +1,3 @@
 export {useVideos, VideoProvider} from "./videoContext";
 export {useAuth, AuthProvider} from "./authContext";
+export {useLikedList, LikesProvider} from "./likesContext";
