+++ src/reducer/likesReducer.js
@@ -1,0 +1,13 @@
+const likesReducer = (state, action) => {
+	switch (action.type) {
+	case "SET_LIKED_LIST":
+		return {
+			...state,
+			likedItems: action.payload,
+		};
+		default:
+			return state;
+	}
+};
+
+export {likesReducer};
